+++ spring/src/main/java/com/example/todo/entity/Task.java
@@ -1,6 +1,6 @@
 package com.example.todo.entity;
 
-import java.time.LocalDateTime;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -8,15 +8,15 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
 @Entity
 @Table(name = "tasks")
-public class Task {
+@EqualsAndHashCode(callSuper = false)
+public class Task extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -44,33 +44,5 @@ public class Task {
   @Column(columnDefinition = "TEXT")
   private String memo;
 
-  @Column(name = "deadline_at")
-  private LocalDateTime deadlineAt;
 
-  @Column(name = "completed_at")
-  private LocalDateTime completedAt;
-
-  @Column(name = "created_at", nullable = false)
-  private LocalDateTime createdAt;
-
-  @Column(name = "updated_at", nullable = false)
-  private LocalDateTime updatedAt;
-
-  /**
-   * @PrePersist JPA（Java Persistence API）のアノテーション。エンティティの新規作成処理が行われる前に呼び出される。
-   */
-  @PrePersist
-  private void onCreate() {
-    LocalDateTime now = LocalDateTime.now();
-    this.setCreatedAt(now);
-    this.setUpdatedAt(now);
-  }
-
-  /**
-   * @PreUpdate JPA（Java Persistence API）のアノテーション。エンティティの更新処理が行われる前に呼び出される。
-   */
-  @PreUpdate
-  private void onUpdate() {
-    this.setUpdatedAt(LocalDateTime.now());
-  }
 }
